removed external utilsfolder, moved queries and mutations scripts to client/src/utils
